add tests for search result
