--- frontend-opal/src/actions/Watchlists.js
@@ -1,10 +1,0 @@
-const getAllData = (data) => ({
-  type: 'GET_ALL_DATA',
-  payload: {
-    data,
-  },
-});
-
-export default {
-  getAllData,
-};
