+++ frontend-opal/src/actions/searchResult.js
