+++ frontend-opal/src/actions/watchlists.js
@@ -1,0 +1,10 @@
+const getWatchlistData = data => ({
+  type: 'GET_WATCHLIST_DATA',
+  payload: {
+    data
+  }
+});
+
+export default {
+  getWatchlistData
+};
