+++ frontend-opal/src/components/SearchResult/Result.js
