+++ frontend-opal/src/components/SearchResult/SearchResult.js
