+++ frontend-opal/src/components/SearchResult/styles/header.js
