+++ frontend-opal/src/components/SearchResult/styles/index.js
