+++ frontend-opal/src/components/SearchResult/styles/result.js
