+++ frontend-opal/src/reducers/searchResult.js
