--- frontend-opal/tests/portfolio/Header.test.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import renderer from 'react-test-renderer';
-import Header from '../../src/components/Header';
-
-describe('<Header />', () => {
-  it('renders correctly', () => {
-    const tree = renderer.create(<Header />).toJSON();
-    expect(tree).toMatchSnapshot();
-  });
-});
